Move user repository to repositories directory to extract abstraction further
