+++ src/user/user.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
-import { UserRepository } from './user.repository';
+import { UserRepository } from '../repositories/user.repository';
 
 @Module({
   controllers: [UserController],
